chg: widget holder to be separate class
