+++ src/view/show_mode/editor/scene_editor.py
@@ -4,7 +4,7 @@ from PySide6.QtWidgets import QDialog, QGridLayout, QTabBar, QTabWidget, QWidget
 
 from model import Scene
 from model.scene import FilterPage
-from view.show_mode.editor.editor_tab_widgets.scene_ui_page_editor_widget import SceneUIPageEditorWidget
+from view.show_mode.editor.editor_tab_widgets.ui_widget_editor.scene_ui_page_editor_widget import SceneUIPageEditorWidget
 
 
 class SceneUIManagerWidget(QTabWidget):
